fix: fix and interate on codesandbox config
